requirements.txt
kornia==0.7.2
matplotlib==3.8.3
numpy==1.26.4
Pillow==10.2.0
PyYAML==6.0.1
scikit_learn==1.5.1
scipy==1.12.0
setproctitle==1.3.3
scikit-image==0.22.0
timm==0.9.16
torchsummary==1.5.1
tqdm==4.66.1
einops==0.7.0
audtorch==0.6.4
tensorboard==2.14.0
pytorch-msssim==1.0.0 
thop
